Ensure that username is lowercase (#109)
My username is set as "Jerry" and on first `tilt up` I hit this error for all but three of the services. 

`a DNS-1123 subdomain must consist of lower case alphanumeric characters, '-' or '.', and must start and end with an alphanumeric character.`

Converting the username to lowercase resolves this issue.